added GLCD lib. symlink it to your arduino/libraries added sketchbook. symlink it to your ~/ (or wherever)
demo program for the board is in sketchbook.
Green LED is speaker for now.

need to debug wireless bug, for now you can comment that out
DO NOT PROGRAM THE BOARD WITH A XBEE IN take the xbee out to program it.
theres an extra xbee socket on the breadboard to hold your radio since youre gonna take it out to program the board